feat: Finzaling the controller, was added some exceptions

+++ src/tasks/tasks.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Delete, Put, Param, Query, Body, HttpStatus } from '@nestjs/common';
+import { Controller, Get, Post, Delete, Put, Param, Query, Body, HttpStatus, ConflictException, NotFoundException } from '@nestjs/common';
 import { TasksService } from './tasks.service';
 import { CreateTaskDto } from 'src/dto/create-task.dto';
 import { UpdateTaskDto } from 'src/dto/update-task.dto';
@@ -8,6 +8,7 @@ import { TaskResponseDto } from 'src/dto/task.response.dto';
 import { GetTaskResponse } from './response/get-task.response';
 import { UpdateTaskResponse } from './response/update-task.response';
 import { DeleteTaskResponse } from './response/delete-task.response';
+import { ExceptionsHandler } from '@nestjs/core/exceptions/exceptions-handler';
 
 @Controller('tasks')
 export class TasksController {
@@ -24,34 +25,48 @@ export class TasksController {
 
     @Get(':id')
     async findOneTask(@Param('id')id:string): Promise<GetTaskResponse>{
+        const task = await this.taskService.findOne(id)
+        if (!task) throw new NotFoundException('Task not found')
         return {
             statusCode: HttpStatus.OK,
             message: 'Task obtained succesfully',
-            data: await this.taskService.findOne(id)
+            data: task
         }
     }
 
     @Post()
     async createTask(@Body() body: CreateTaskDto): Promise<CreateTaskResponse>{
-        return {
-            statusCode: HttpStatus.CREATED,
-            message: 'Task created succesfully',
-            data: await this.taskService.create(body)
+        try {
+            return {
+                statusCode: HttpStatus.CREATED,
+                message: 'Task created succesfully',
+                data: await this.taskService.create(body)
+            }
+        } catch (error) {
+            if (error.code === 11000){
+                throw new ConflictException('Task already exist')
+            }
+            throw new error
         }
     }
 
     @Put(':id')
     async updateTask(@Param('id')id:string, @Body() body: UpdateTaskDto): Promise<UpdateTaskResponse>{
+        const updatedTask = await this.taskService.update(id, body)
+        if (!updatedTask) throw new NotFoundException('Task not found')
         return {
             statusCode: HttpStatus.OK,
             message: 'Task updated succesfully',
-            data: await this.taskService.update(id, body)
+            data: updatedTask
         }
     }
 
     @Delete(':id')
     async deleteTask(@Param('id') id:string): Promise<DeleteTaskResponse>{
-        this.taskService.delete(id)
+        if (id.length != 24) throw new NotFoundException('The id must have a specific lenght.')
+        const task = await this.taskService.delete(id)
+        if (!task) throw new NotFoundException('Task not found')
+
         return {
             statusCode: HttpStatus.OK,
             message: 'Task succesfully deleted.',
